Create component for task parameters & clean up
