+++ hat/assets/js/apps/Iaso/domains/tasks/components/TaskModal.tsx
@@ -1,15 +1,6 @@
 import React, { FunctionComponent, useEffect, useRef, useState } from 'react';
 import RemoveRedEyeIcon from '@mui/icons-material/RemoveRedEye';
-import {
-    Box,
-    CircularProgress,
-    Grid,
-    IconButton,
-    Table,
-    TableBody,
-    TableCell,
-    TableRow,
-} from '@mui/material';
+import { Box, CircularProgress, Grid, IconButton } from '@mui/material';
 import { AlertModal, makeFullModal, useSafeIntl } from 'bluesquare-components';
 import { useQueryClient } from 'react-query';
 import WidgetPaper from 'Iaso/components/papers/WidgetPaperComponent';
@@ -17,19 +8,8 @@ import { TaskBaseInfo } from 'Iaso/domains/tasks/components/TaskBaseInfo';
 import { TaskLogMessages } from 'Iaso/domains/tasks/components/TaskLogMessages';
 import { useGetLogs, useGetTask } from 'Iaso/domains/tasks/hooks/api';
 import { Task } from 'Iaso/domains/tasks/types';
-import { SxStyles } from 'Iaso/types/general';
 import MESSAGES from '../messages';
-
-type TaskParams = {
-    label: string;
-    value: any;
-};
-
-const styles: SxStyles = {
-    table: {
-        tableLayout: 'fixed',
-    },
-};
+import { TaskParameters, TaskParams } from './TaskParameters';
 
 export type Props = {
     task: Task<any>;
@@ -75,22 +55,7 @@ const TaskModal: FunctionComponent<Props> = ({
                 </Grid>
                 {taskParams && taskParams.length > 0 ? (
                     <Grid item xs={6}>
-                        <WidgetPaper title={formatMessage(MESSAGES.params)}>
-                            <Table
-                                size={'small'}
-                                data-test="task-base-info"
-                                sx={styles.table}
-                            >
-                                <TableBody>
-                                    {taskParams.map(p => (
-                                        <TableRow key={p.label}>
-                                            <TableCell>{p.label}</TableCell>
-                                            <TableCell>{p.value}</TableCell>
-                                        </TableRow>
-                                    ))}
-                                </TableBody>
-                            </Table>
-                        </WidgetPaper>
+                        <TaskParameters taskParams={taskParams} />
                     </Grid>
                 ) : null}
             </Grid>
